refactor: move cmds/
