+++ src/kernel/cmd.cpp
@@ -1,8 +1,6 @@
 #include "cmd.hpp"
 
-#include "board/mailbox.hpp"
 #include "board/mini-uart.hpp"
-#include "board/pm.hpp"
 #include "string.hpp"
 
 int help_idx = 0;
@@ -29,28 +27,6 @@ const Cmd cmds[] = {
     },
 };
 constexpr int ncmd = sizeof(cmds) / sizeof(cmds[0]);
-
-void cmd_help() {
-  for (int i = 0; i < ncmd; i++) {
-    auto cmd = &cmds[i];
-    mini_uart_printf("%s\t%s\n", cmd->name(), cmd->help());
-  }
-}
-
-void cmd_hello() {
-  mini_uart_puts("Hello World!\n");
-}
-
-void cmd_hwinfo() {
-  // it should be 0xa020d3 for rpi3 b+
-  mini_uart_printf("Board revision :\t0x%08X\n", get_board_revision());
-  mini_uart_printf("ARM memory base:\t0x%08X\n", get_arm_memory(0));
-  mini_uart_printf("ARM memory size:\t0x%08X\n", get_arm_memory(1));
-}
-
-void cmd_reboot() {
-  reboot();
-}
 
 void runcmd(const char* buf, int len) {
   const Cmd* cmd = nullptr;
