+++ src/kernel/cmds/reboot.cpp
@@ -1,0 +1,6 @@
+#include "board/pm.hpp"
+#include "cmd.hpp"
+
+void cmd_reboot() {
+  reboot();
+}
